[#noissue] Refactor datatype of ServiceType in Server from short to int

+++ agent-module/plugins-loader/src/main/java/com/navercorp/pinpoint/loader/service/DefaultServiceTypeRegistryService.java
@@ -36,7 +36,7 @@ public class DefaultServiceTypeRegistryService implements ServiceTypeRegistrySer
     }
 
     @Override
-    public ServiceType findServiceType(short serviceType) {
+    public ServiceType findServiceType(int serviceType) {
         return serviceTypeLocator.findServiceType(serviceType);
     }
