+++ web/src/main/java/com/navercorp/pinpoint/web/component/DefaultAnnotationKeyMatcherService.java
@@ -37,7 +37,7 @@ public class DefaultAnnotationKeyMatcherService implements AnnotationKeyMatcherS
     }
 
     @Override
-    public AnnotationKeyMatcher findAnnotationKeyMatcher(short serviceType) {
+    public AnnotationKeyMatcher findAnnotationKeyMatcher(int serviceType) {
         return annotationKeyMatcherLocator.findAnnotationKeyMatcher(serviceType);
     }
 }
